+++ addons/ui_react/editor_plugin/ui_react_dock.gd
@@ -530,6 +530,7 @@ func refresh() -> void:
 
 	if root:
 		_issues_all.append_array(UiReactValidatorService.validate_wiring_under_root(root))
+		_issues_all.append_array(UiReactValidatorService.validate_computed_under_root(root))
 
 	_issues_all.append_array(UiReactUnusedStateService.build_issues(_dock_actions.resolve_output_dir(), root))